backend/requirements.txt
Flask==3.0.3
mysql-connector==2.2.9
bcrypt==4.2.0
pyjwt==2.9.0
datetime==5.5
